Revert "chore: update modal and add confirm"
This reverts commit 1673f42e346d97639f1d218ac3d54e5ccb000751.
