+++ src/components/Modal/index.js
@@ -6,18 +6,20 @@ import styles from './style.module.css'
 
 const Modal = ({
   children,
+  close,
   show,
 }) => (
   <div
+    onClick={close}
     className={classNames(styles.bgModal, {
       [styles.show]: show,
     })}
   >
-    <div
-      className={classNames(styles.modal, {
-        [styles.show]: show,
-      })}
-    >
+  <div
+    className={classNames(styles.modal, {
+      [styles.show]: show,
+    })}
+  >
       {children}
     </div>
    </div>
@@ -25,6 +27,7 @@ const Modal = ({
 
 Modal.propTypes = {
   children: PropTypes.node.isRequired,
+  close: PropTypes.func.isRequired,
   show: PropTypes.bool.isRequired,
 }
